Implements Cascaded Shadow Maps for directional lights
Directional lights now support three shadow maps, which must have their
dimensions set properly by the user via PerspectiveCamera.getWorldSpaceSubFrusta and directionalLight.updateCascades. It assumes the nearest map is the first,
etc... and that the first map represents the closest 0.01 of the range, the second is up to 0.1 of the range, and the last is the rest.

Ensures "smooth" (linear) blending between levels of the cascade as well.

Tested with physical, phong, and lambert materials, materials now require
the near and far clipping plane values to be passed in.