Improves CSM work
Near and Far clipping planes are no longer stored in the material,
but are grabbed from the camera when refreshing uniforms.

This is cleaner for a few reasons: less work the user needs to do, and less weird state to maintain mainly.

As a consequence, the unneeded cruft has been removed.