Adds support for direct light maps
The previous "light map" property of materials has now been split into "indirect light map" and an "indirect light map" with distinct intensities.

As part of this, lights should all support an "isDynamicLight" property, which indicates whether the light should contribute direct light to materials with direct light maps. If a light is marked as a dynamic light, it will do its full lighting (and shadowing calculation for all materials). If a light is instead marked as static, it should only contribute specular lighting, not shadows or diffuse lighting.