Change GLTFExporter to use standard uv attribute
For some reason, the existing GLTF export duplicates UVs for the AO Map, if it is present, which causes many GLTF viewers and iOS Quick View to not actually render the occlusion maps due to limited support for multiple uv maps.

Since using the same uv map for both the occlusion and the roughness metalness makes perfect sense we simply remove the specification of the uvmap for the occlusion texture.